Refactor DialCounter for a cleaner solution
The value is stored without adjustment (i.e. not the real dial value, but the effective difference since starting), since we're only interested in multiples of 100. Checking stopping at 0 is then done with a simple %100 approach (doesn't matter which multiple, or if negative).

Partial rotations past zero are detected by checking if the effective value moves to a different 100 "block" (i.e. /100 before and after is different). An extra case is needed for when moving across effective value 0 (basically for when going from "negative" zero to "positive" zero or vice versa).